pyproject.toml
[project]
name = ""
version = ""
description = ""
authors = [
    {name = "Yanda Huang", email = "realyanda@hey.com"},
]
dependencies = [
    "soco",
    "pillow",
    "requests",
    "lyricsgenius",
    "toml",
    "fastapi",
    "uvicorn[standard]",
]
requires-python = ">=3.8"
license = {text = "MIT"}

[project.urls]
homepage = ""

[tool]
[tool.pdm]

[build-system]
requires = ["pdm-pep517"]
build-backend = "pdm.pep517.api"
